+++ src/ApiComponents/Form/FormCreatePayment/index.tsx
@@ -1,0 +1,2 @@
+import FormCreatePayment from "./FormCreatePayment";
+export default FormCreatePayment;
