+++ src/ApiComponents/Grid/GridPayment/index.tsx
@@ -1,0 +1,2 @@
+import GridPayment from "./GridPayment";
+export default GridPayment;
