--- src/ApiComponents/Grid/index.tsx
@@ -1,2 +1,0 @@
-import Grid from "./Grid";
-export default Grid;
