+++ src/Components/NavTop/NavTop.tsx
@@ -44,6 +44,7 @@ const Container = styled.div<IContainer>`
     align-items: center;
     height: ${props => props.navHeight}px;
     background-color: ${props => props.theme.navTopColor};
+    box-shadow: 0 1px 2px 1px rgba(0,0,0,.22);
 `;
 const Wrapper = styled.div`
     padding: 20px;
